requirements.txt
fastapi
uvicorn
torch
torchvision
diffusers
transformers
accelerate
safetensors
controlnet-aux
opencv-python
Pillow
